requirements.txt
build==1.2.1
click==8.1.7
dagu-cron==0.1.0
importlib_metadata==8.0.0
packaging==24.1
pyaml==24.4.0
pyproject_hooks==1.1.0
PyYAML==6.0.1
tomli==2.0.1
zipp==3.19.2
